chore: 背景色の変更と軽微な修正
